+++ app/src/helpers/helpers.ts
@@ -284,6 +284,16 @@ export function openExternal(
   return shell.openExternal(url, options);
 }
 
+// Copy-pastaed as unable to get imports to work in preload.
+// If modifying, update also app/src/preload.ts
+export function isWayland(): boolean {
+  return (
+    isLinux() &&
+    (Boolean(process.env.WAYLAND_DISPLAY) ||
+      process.env.XDG_SESSION_TYPE === 'wayland')
+  );
+}
+
 export function removeUserAgentSpecifics(
   userAgentFallback: string,
   appName: string,